chore: move lightclient message to types
